+++ Sources/RVS_Generic_Swift_Toolbox/Extensions/RVS_Bundle_Extensions.swift
@@ -19,7 +19,7 @@ CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFT
 
 The Great Rift Valley Software Company: https://riftvalleysoftware.com
  
- Version: 1.6.7
+ Version: 1.7.0
 */
 import Foundation   // Required for Bundle